sqlite_to_postgres/requirements.txt
environs==8.0.0
psycopg2-binary==2.9.6